Move dog optimizer to contrib
